+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2024-06-13
+
+### Changed
+
+- Optimize filtering time
+- Reorganize buttons filters as components
 
 ## [0.3.0] - 2024-05-26
 
@@ -33,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Select working directory
 - Show analysis: files type and name formats
 
-[Unreleased]: https://github.com/vikian050194/tacktick-client/compare/v0.2.0...HEAD
-[0.2.0]: https://github.com/vikian050194/tacktick-client/compare/v0.1.0...v0.2.0
-[0.1.0]: https://github.com/vikian050194/tacktick-client/releases/tag/v0.1.0
+[Unreleased]: https://github.com/vikian050194/tart/compare/v0.3.1...HEAD
+[0.3.1]: https://github.com/vikian050194/tart/compare/v0.3.0...v0.3.1
+[0.3.0]: https://github.com/vikian050194/tart/compare/v0.2.0...v0.3.0
+[0.2.0]: https://github.com/vikian050194/tart/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/vikian050194/tart/releases/tag/v0.1.0